+++ test/OpenTelemetry.Instrumentation.Wcf.Tests/TelemetryDispatchMessageInspectorTests.netfx.cs
@@ -63,10 +63,7 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
         this.serviceHost = createdHost;
     }
 
-    public void Dispose()
-    {
-        this.serviceHost?.Close();
-    }
+    public void Dispose() => this.serviceHost?.Close();
 
     [Theory]
     [InlineData(true, false)]
@@ -88,7 +85,7 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
 
         using var activityListener = new ActivityListener
         {
-            ShouldListenTo = activitySource => true,
+            ShouldListenTo = _ => true,
             ActivityStopped = stoppedActivities.Add,
         };
 
@@ -146,15 +143,7 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
         }
         finally
         {
-            if (client.State == CommunicationState.Faulted)
-            {
-                client.Abort();
-            }
-            else
-            {
-                client.Close();
-            }
-
+            client.AbortOrClose();
             tracerProvider?.Shutdown();
             tracerProvider?.Dispose();
 
@@ -164,34 +153,24 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
         if (instrument && !filter)
         {
             Assert.NotEmpty(stoppedActivities);
-            Assert.Single(stoppedActivities);
-
-            var activity = stoppedActivities[0];
+            var activity = Assert.Single(stoppedActivities);
 
             if (emptyOrNullAction)
             {
                 Assert.Equal(WcfInstrumentationActivitySource.IncomingRequestActivityName, activity.DisplayName);
-                Assert.Equal("ExecuteWithEmptyActionName", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.RpcMethodTag).Value);
-                Assert.Equal("http://opentelemetry.io/Service/ExecuteWithEmptyActionNameResponse", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.SoapReplyActionTag).Value);
+                Assert.Equal("ExecuteWithEmptyActionName", activity.TagObjects.FirstOrDefault(t => t.Key == SemanticConventions.AttributeRpcMethod).Value);
+                Assert.Equal("http://opentelemetry.io/Service/ExecuteWithEmptyActionNameResponse", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.AttributeSoapReplyAction).Value);
             }
             else
             {
                 Assert.Equal("http://opentelemetry.io/Service/Execute", activity.DisplayName);
-                Assert.Equal("Execute", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.RpcMethodTag).Value);
-                Assert.Equal("http://opentelemetry.io/Service/ExecuteResponse", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.SoapReplyActionTag).Value);
+                Assert.Equal("Execute", activity.TagObjects.FirstOrDefault(t => t.Key == SemanticConventions.AttributeRpcMethod).Value);
+                Assert.Equal("http://opentelemetry.io/Service/ExecuteResponse", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.AttributeSoapReplyAction).Value);
             }
-
-            Assert.Equal(WcfInstrumentationActivitySource.IncomingRequestActivityName, activity.OperationName);
-            Assert.Equal(WcfInstrumentationConstants.WcfSystemValue, activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.RpcSystemTag).Value);
-            Assert.Equal("http://opentelemetry.io/Service", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.RpcServiceTag).Value);
-            Assert.Equal(this.serviceBaseUri.Host, activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.NetHostNameTag).Value);
-            Assert.Equal(this.serviceBaseUri.Port, activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.NetHostPortTag).Value);
-            Assert.Equal("net.tcp", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.WcfChannelSchemeTag).Value);
-            Assert.Equal("/Service", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.WcfChannelPathTag).Value);
 
             if (includeVersion)
             {
-                Assert.Equal("Soap12 (http://www.w3.org/2003/05/soap-envelope) Addressing10 (http://www.w3.org/2005/08/addressing)", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.SoapMessageVersionTag).Value);
+                Assert.Equal("Soap12 (http://www.w3.org/2003/05/soap-envelope) Addressing10 (http://www.w3.org/2005/08/addressing)", activity.TagObjects.FirstOrDefault(t => t.Key == WcfInstrumentationConstants.AttributeSoapMessageVersion).Value);
             }
 
             if (enrich && !enrichmentException)
@@ -199,6 +178,8 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
                 Assert.Equal(WcfEnrichEventNames.AfterReceiveRequest, activity.TagObjects.Single(t => t.Key == "server.afterreceiverequest").Value);
                 Assert.Equal(WcfEnrichEventNames.BeforeSendReply, activity.TagObjects.Single(t => t.Key == "server.beforesendreply").Value);
             }
+
+            WcfTestHelpers.AssertIncomingRequestActivityCommon(activity, this.serviceBaseUri);
         }
         else
         {
@@ -226,12 +207,12 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
         List<Exception> recordedExceptions = [];
         using var activityListener = new ActivityListener
         {
-            ShouldListenTo = activitySource => true,
+            ShouldListenTo = _ => true,
             ActivityStarted = startedActivities.Add,
             ActivityStopped = stoppedActivities.Add,
         };
 
-        activityListener.ExceptionRecorder += (Activity activity, Exception ex, ref TagList tags) =>
+        activityListener.ExceptionRecorder += (activity, ex, ref tags) =>
         {
             recordedExceptions.Add(ex);
         };
@@ -286,15 +267,7 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
             startedActivities[0].AddTag(nameof(triggerException), triggerException);
             startedActivities[0].AddTag(nameof(runAsync), runAsync);
 
-            if (client.State == CommunicationState.Faulted)
-            {
-                client.Abort();
-            }
-            else
-            {
-                client.Close();
-            }
-
+            client.AbortOrClose();
             tracerProvider?.Shutdown();
             tracerProvider?.Dispose();
 
@@ -306,7 +279,7 @@ public class TelemetryDispatchMessageInspectorTests : IDisposable
 
         if (recordException && triggerException)
         {
-            Assert.Collection(recordedExceptions, e => Assert.IsType<Exception>(e));
+            Assert.All(recordedExceptions, e => Assert.IsType<Exception>(e));
         }
         else
         {